Update double_entry_check.R
I revised my code again because there is a bug in the previous version. I removed all the breaks and used the return(NULL) statement instead to stop the functions if certain conditions are not met. This change can get rid of the build-in errors in R and the use of try() with (silent = TRUE).